% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanUnique.R
\name{meanUnique}
\alias{meanUnique}
\title{Calulate mean or take unique elements of a vector}
\usage{
meanUnique(x)
}
\arguments{
\item{x}{a vector}
}
\value{
If x is numeric, returns the mean. Otherwise, returns the unique elements of x.
}
\description{
A function that calulate mean or take unique elements of a vector.
}
\examples{
x <- 1:5
meanUnique(x)
x=c("a","a","b")
meanUnique(x)
}
