% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRank.R
\name{pRank}
\alias{pRank}
\title{Percentile rank transformation of the data}
\usage{
pRank(x, xSample)
}
\arguments{
\item{x}{A data frame containing the pooled data from fcs files. Each row is
a cell, each column is a marker.}

\item{xSample}{A vector specifying which sample each cell belongs to. Length
must equal to nrow(x).}
}
\value{
Returns data frame containing rank transformed data.
}
\description{
A function that performs the rank transformation of the data.
}
\examples{
x <- pRank(x=iris[,1:4],xSample=iris$Species)
}
