% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.ub.average.R
\name{rank.ub.average}
\alias{rank.ub.average}
\title{Percentile rank transformation of a vector (ties.method = "average")}
\usage{
rank.ub.average(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns the percentile rank of each element.
}
\description{
A function that performs the rank transformation of a vector (ties.method = "average").
}
\examples{
rank.ub.average(1:10)
}
