% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.ub.random.R
\name{rank.ub.random}
\alias{rank.ub.random}
\title{Percentile rank transformation of a vector (ties.method = "random")}
\usage{
rank.ub.random(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Returns the percentile rank of each element.
}
\description{
A function that performs the rank transformation of a vector (ties.method = "random").
}
\examples{
rank.ub.average(1:10)
}
